move injected plugins into plugins directory
