+++ packages/houdini/src/codegen/generators/runtime/runtime.test.ts
@@ -42,7 +42,7 @@ test('updates the list of plugin-specified client plugins', async function () {
 
 	// open up the index file
 	const fileContents = await fs.readFile(
-		path.join(config.runtimeDirectory, 'client', 'injectedPlugins.js')
+		path.join(config.runtimeDirectory, 'client', 'plugins', 'injectedPlugins.js')
 	)
 	expect(fileContents).toBeTruthy()
 
@@ -111,7 +111,7 @@ test('passing null as client plugin config generates a reference, not a function
 
 	// open up the index file
 	const fileContents = await fs.readFile(
-		path.join(config.runtimeDirectory, 'client', 'injectedPlugins.js')
+		path.join(config.runtimeDirectory, 'client', 'plugins', 'injectedPlugins.js')
 	)
 	expect(fileContents).toBeTruthy()
