+++ packages/houdini/src/runtime/client/plugins/injectedPlugins.ts
@@ -1,7 +1,7 @@
 // the actual contents of this file will be overwritten by the runtime generator
 // to include imports from plugins so that they can register middlewares as
 // part of the generation pipeline
-import { ClientPlugin } from './documentObserver'
+import { ClientPlugin } from '../documentObserver'
 
 const plugins: ClientPlugin[] = []
